requirements.txt
opencv-contrib-python>=4.8.0
numpy>=1.24.0
tqdm>=4.65.0
ffmpeg-python>=0.2.0